+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ### 5.2.1 (TBD)
 - Fixed regression in parser where group length of VR UN was parsed wrong (#1941)
 - UseFellowOakDicom registers the default services only if there are not yet some services registered (#1929)
+- update to DICOM Standard 2025a
 
 ### 5.2.0 (2025-02-03)
 - refactor the parser to make it better maintainable